requirements.txt
pandas
numpy
langchain
langchain-openai
langchain_core
langchain-community
langchainhub
openai
ragas
python-dotenv
pdfplumber
faiss_cpu
ipykernel
pypdf
websocket-client
text_generation
chainlit
cohere
tiktoken
